feat: add publish new verion button in brand app
